src/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
discord==1.7.3
discord.py==1.7.3
idna==3.2
loguru==0.5.3
multidict==5.1.0
mysql-connector-python==8.0.26
protobuf==3.18.0
pycparser==2.20
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
ssdeep==3.4
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
